+++ go.mod
@@ -3,10 +3,10 @@ module github.com/google/osv-scanner
 go 1.21.12
 
 require (
-	deps.dev/api/v3 v3.0.0-20240711010811-c5a1406b0470
-	deps.dev/util/maven v0.0.0-20240711010811-c5a1406b0470
-	deps.dev/util/resolve v0.0.0-20240711010811-c5a1406b0470
-	deps.dev/util/semver v0.0.0-20240711010811-c5a1406b0470
+	deps.dev/api/v3 v3.0.0-20240730004939-e80e6658c33b
+	deps.dev/util/maven v0.0.0-20240730004939-e80e6658c33b
+	deps.dev/util/resolve v0.0.0-20240730004939-e80e6658c33b
+	deps.dev/util/semver v0.0.0-20240730004939-e80e6658c33b
 	github.com/BurntSushi/toml v1.4.0
 	github.com/CycloneDX/cyclonedx-go v0.9.0
 	github.com/charmbracelet/bubbles v0.18.0
