--- internal/resolution/manifest/fixtures/maven/pom.xml
@@ -1,10 +1,0 @@
-<!-- For testing parent relative path -->
-<project>
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>org.test</groupId>
-  <artifactId>test</artifactId>
-  <version>1.0.0</version>
-
-</project>
